Use just one error to be thrown during mutation and move namespace
