+++ tests/Mutant/Generator/MutationsGeneratorTest.php
@@ -17,7 +17,7 @@ use Infection\Mutator\Arithmetic\Decrement;
 use Infection\Mutator\Arithmetic\Plus;
 use Infection\Mutator\Boolean\TrueValue;
 use Infection\Mutator\FunctionSignature\PublicVisibility;
-use Infection\Mutator\Util\Exception\MutatorException;
+use Infection\Exception\MutatorException;
 use Infection\Mutator\Util\MutatorConfig;
 use Infection\TestFramework\Coverage\CodeCoverageData;
 use Infection\Tests\Fixtures\Files\Mutation\OneFile\OneFile;
